remove code duplication in collection converters
