+++ src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Converters/Collection/IAsyncEnumerableOfTConverter.cs
@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 namespace System.Text.Json.Serialization.Converters
 {
     internal sealed class IAsyncEnumerableOfTConverter<TAsyncEnumerable, TElement>
-        : IEnumerableDefaultConverter<TAsyncEnumerable, TElement>
+        : JsonCollectionConverter<TAsyncEnumerable, TElement>
         where TAsyncEnumerable : IAsyncEnumerable<TElement>
     {
         internal override bool OnTryRead(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options, ref ReadStack state, out TAsyncEnumerable value)
